Addressed Review feedback, extracted EmptyCursor to a own class to reduce duplicated code
