+++ app/core/src/main/java/com/fsck/k9/helper/EmptyCursor.java
@@ -1,0 +1,54 @@
+package com.fsck.k9.helper;
+
+import android.database.AbstractCursor;
+
+
+/**
+ * A dummy class that provides a empty cursor
+ */
+public class EmptyCursor extends AbstractCursor {
+    @Override
+    public int getCount() {
+        return 0;
+    }
+
+    @Override
+    public String[] getColumnNames() {
+        return new String[0];
+    }
+
+    @Override
+    public String getString(int column) {
+        return null;
+    }
+
+    @Override
+    public short getShort(int column) {
+        return 0;
+    }
+
+    @Override
+    public int getInt(int column) {
+        return 0;
+    }
+
+    @Override
+    public long getLong(int column) {
+        return 0;
+    }
+
+    @Override
+    public float getFloat(int column) {
+        return 0;
+    }
+
+    @Override
+    public double getDouble(int column) {
+        return 0;
+    }
+
+    @Override
+    public boolean isNull(int column) {
+        return false;
+    }
+}
